Bug 34643: Split CircConfirmItemParts for self-checkout and self-checkin
Many libraries wish to be able to allow self-check for multi-part items while requesting librarians to verify the parts when checking out/in at the desk. It makes sense to split CircConfirmItemParts into three prefs such that the feature can be enabled/disabled for SCO and SCI independently.

Test Plan:
1) Apply this patch
2) Enable SCO and SCI
3) Set CircConfirmItemPartsSCO to "Don't allow"
4) Attempt to checkout out an item with materials via SCO, it should
   fail
5) Set CircConfirmItemPartsSCO to "Allow"
6) Repeat step 4, it should succeed
7) Set CircConfirmItemPartsSCI to "Don't allow"
8) Attempt to check in the item via SCI, it should fail
9) Attempt to check in the item via SCO, it should fail
9) Set CircConfirmItemPartsSCI to "Allow"
10) Attempt to check in the item via SCO and SCI, it should succeed!

Signed-off-by: AFHDubCoLib <andrewfh@dubcolib.org>