Bug 34784: Implement feature in staff interface
Some libraries would like to auto-populate item callnumbers for items on a record based on the system preference "itemcallnumber". Only items with empty callnumbers would affected. This would add a button "Populate empty item callnumbers" on catalogue/detail.pl, as well as a "Populate callnumber" button for each line in the holdings table to set the callnumber for that specific item.

Test Plan:
1) Set the systempreference to something ( "245a" will make the
   callnumber the record title for testing purposes )
2) Apply this patch set
3) Restart all the things!
4) Browse to a record details page, note no new buttons
5) Enable the new syspref EnablePopulateCallnumbers
5) Reload the record details, note the new "Populate callnumbers" button
   in the cataloging toolbar
7) Add one or more items with no callnumber to the record
8) Test the "Populate callnumbers" button
9) Add another item with no callnumber to the record
10) Test the "Populate callnumber" button for that single item

Signed-off-by: David Nind <david@davidnind.com>