Bug 35603: Add plugin feature to allow plugins to warn if any external library modules needed are missing
Some plugins require extra modules not included with Koha, but there is no simple way to let a user know when installing a plugin.

This patch allows a new metadata to be added that is a list of hashes
with the key module. The hash is used for future expansion to allow new
critera to be added such as minimum and maximum module version.

Test Plan:
1) Install any plugin
2) In the plugin metadata, add the following key:
libs => [ { module => 'Business::Barcode::EAN13' } ],
3) Apply this patch
4) Restart all the things!
5) Browse to plugins-home.pl, note the error listing the missing module
6) Install Business::Barcode::EAN13 from cpan
7) Reload the page
8) Note the error is gone!

Signed-off-by: Martin Renvoize <martin.renvoize@ptfs-europe.com>