Bug 40719: Explicitly disable off RELATIVE file paths for plugins for user-entered templates
In bug 39184 we set EVAL_PERL and ABSOLUTE to 0, which is the default. RELATIVE also defaults to 0. For security we should probably specify that as well in case the default is changed somehow.

Test Plan:
1) Apply this patch
2) Generate some notices
3) No change should be noted!

Signed-off-by: noah <noah@inlibro.com>
Signed-off-by: David Cook <dcook@prosentient.com.au>
Signed-off-by: Lucas Gass <lucas@bywatersolutions.com>
(cherry picked from commit bd85f990dac046edd4d33ed6dbae7b815ec87b3e)
Signed-off-by: Jacob O'Mara <jacob.omara@openfifth.co.uk>